Move strict optional state to an instance attribute for faster access
