+++ mypy/join.py
@@ -17,7 +17,7 @@ from mypy.subtypes import (
 )
 from mypy.nodes import INVARIANT, COVARIANT, CONTRAVARIANT
 import mypy.typeops
-from mypy import state
+from mypy.state import state
 
 
 class InstanceJoiner:
